Refactor polyfit_dict.py to polyfit_lookup.py - makes meaning clearer
