+++ ghostdr/ghost/lookups/polyfit_lookup.py
